+++ app/src/main/java/nextstep/shoppingcart/component/topbar/ShoppingDetailTopBar.kt
@@ -1,37 +1,35 @@
-package nextstep.shoppingcart.component
+package nextstep.shoppingcart.component.topbar
 
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ShoppingCart
+import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 
 @Composable
-fun ShoppingListTopBar(
+fun ShoppingDetailTopBar(
     title: String,
-    onClickCart : () -> Unit,
+    onClickBack: () -> Unit,
     modifier: Modifier = Modifier
 ) {
     ShoppingTopBar(
         modifier = modifier,
         title = title,
+        titleTextAlign = TextAlign.Start,
         leftActionButtons = {
-            Spacer(modifier = Modifier.size(48.dp))
-        },
-        rightActionButtons = {
             IconButton(
-                onClick = onClickCart
+                onClick = onClickBack
             ) {
                 Icon(
                     modifier = Modifier
                         .size(20.dp),
-                    imageVector = Icons.Default.ShoppingCart,
+                    imageVector = Icons.Default.ArrowBack,
                     contentDescription = "ShoppingCart"
                 )
             }
@@ -39,12 +37,12 @@ fun ShoppingListTopBar(
     )
 }
 
-@Preview(name = "ShoppingListTopBar", showBackground = true)
+@Preview(name = "ShoppingDetailTopBar", showBackground = true)
 @Composable
 private fun Preview1() {
-    ShoppingListTopBar(
+    ShoppingDetailTopBar(
         modifier = Modifier.fillMaxWidth(),
         title = "상품 목록",
-        onClickCart = {}
+        onClickBack = {}
     )
 }
