+++ app/src/main/java/nextstep/shoppingcart/component/topbar/ShoppingListTopBar.kt
@@ -1,35 +1,37 @@
-package nextstep.shoppingcart.component
+package nextstep.shoppingcart.component.topbar
 
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.material.icons.filled.ShoppingCart
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 
 @Composable
-fun ShoppingDetailTopBar(
+fun ShoppingListTopBar(
     title: String,
-    onClickBack: () -> Unit,
+    onClickCart : () -> Unit,
     modifier: Modifier = Modifier
 ) {
     ShoppingTopBar(
         modifier = modifier,
         title = title,
-        titleTextAlign = TextAlign.Start,
         leftActionButtons = {
+            Spacer(modifier = Modifier.size(48.dp))
+        },
+        rightActionButtons = {
             IconButton(
-                onClick = onClickBack
+                onClick = onClickCart
             ) {
                 Icon(
                     modifier = Modifier
                         .size(20.dp),
-                    imageVector = Icons.Default.ArrowBack,
+                    imageVector = Icons.Default.ShoppingCart,
                     contentDescription = "ShoppingCart"
                 )
             }
@@ -37,12 +39,12 @@ fun ShoppingDetailTopBar(
     )
 }
 
-@Preview(name = "ShoppingDetailTopBar", showBackground = true)
+@Preview(name = "ShoppingListTopBar", showBackground = true)
 @Composable
 private fun Preview1() {
-    ShoppingDetailTopBar(
+    ShoppingListTopBar(
         modifier = Modifier.fillMaxWidth(),
         title = "상품 목록",
-        onClickBack = {}
+        onClickCart = {}
     )
 }
